fix test json, move shared util to common
