+++ app/src/js/server/bot.ts
@@ -5,8 +5,8 @@ import uuid4 from 'uuid/v4'
 import { ADD_LABELS, AddLabelsAction, BaseAction } from '../action/types'
 import { configureStore, ReduxStore } from '../common/configure_store'
 import { ShapeTypeName } from '../common/types'
+import { index2str } from '../common/util'
 import { PolygonType, RectType, State } from '../functional/types'
-import { index2str } from '../shared/util'
 import { ItemExport } from './bdd_types'
 import Logger from './logger'
 import { ModelInterface } from './model_interface'
