+++ app/src/js/server/path.ts
@@ -2,7 +2,7 @@ import moment from 'moment'
 import * as os from 'os'
 import * as path from 'path'
 import { sprintf } from 'sprintf-js'
-import { index2str } from '../shared/util'
+import { index2str } from '../common/util'
 import { BotData } from './types'
 
 /**
