+++ app/src/test/server/s3_storage.test.ts
@@ -1,9 +1,9 @@
 import AWS from 'aws-sdk'
 import * as path from 'path'
 import { sprintf } from 'sprintf-js'
+import { index2str } from '../../js/common/util'
 import { getProjectKey, getTaskKey, hostname, now } from '../../js/server/path'
 import { S3Storage } from '../../js/server/s3_storage'
-import { index2str } from '../../js/shared/util'
 
 export const s3 = new AWS.S3()
 const bucketRegion = 'us-west-2'
